+++ test/e2e/cypress/integration/commands.spec.js
@@ -42,10 +42,6 @@ context("Commands", function () {
   context("NTLM config", function () {
     const httpHost = "http://localhost:5000";
 
-    beforeEach("Reset NTLM config", function () {
-      cy.ntlmReset();
-    });
-
     it("cy.ntlm with invalid ntlmHosts shall not succeed", function () {
       expectFail("Invalid ntlmHosts, must be an array.");
       cy.ntlm({}, "nis\\se", "manpower", "mpatst");
@@ -86,11 +82,11 @@ context("Commands", function () {
     });
   });
 
-  context("NTLM SSO config", function () {
-    const httpHost = "http://localhost:5000";
-
-    beforeEach("Reset NTLM config", function () {
-      //cy.ntlmReset();
+  context("NTLM SSO config win32", function () {
+    before("Check platform", function () {
+      if (Cypress.platform !== "win32") {
+        this.skip();
+      }
     });
 
     it("cy.ntlmSso with protocol in ntlmHosts shall not succeed", function () {
@@ -124,45 +120,29 @@ context("Commands", function () {
       cy.ntlmSso(undefined);
     });
 
-    it.only("cy.ntlmSso with on non windows shall not succeed", function () {
+    it("cy.ntlmSso on non windows shall not succeed", function () {
       if (Cypress.platform === "win32") {
         this.skip();
       }
-      /*
       expectFail(
-        "Body: SSO config parse error. SSO is not supported on this platform. Only Windows OSs are supported."
+        "SSO is not supported on this platform. Only Windows OSs are supported."
       );
-      */
-
-      cy.log("Request without log");
-      cy.request({
-        method: "GET",
-        url: "https://jsonplaceholder.cypress.io/todos/2",
-        log: false,
-      });
-      /*
-      cy.request({
-        method: "POST",
-        url: "https://this.does.not.exxizzt",
-        log: false,
-        body: { muu: "more" },
-        headers: { "x-my-header": "yess" },
-      });
-*/
-      cy.log("Request with log");
-      cy.request({
-        method: "GET",
-        url: "https://jsonplaceholder.cypress.io/todos/2",
-      });
-
-      cy.request({
-        method: "POST",
-        url: "http://www.mocky.io/v2/5ec993803000009700a6ce1f",
-        log: false,
-      });
-
       cy.ntlmSso(["nisse.com", "google.com"]);
-      //cy.get("body").should("have.value", "dummy");
+    });
+  });
+
+  context("NTLM SSO config non win32", function () {
+    before("Check platform", function () {
+      if (Cypress.platform === "win32") {
+        this.skip();
+      }
+    });
+
+    it("cy.ntlmSso on non windows shall not succeed", function () {
+      expectFail(
+        "SSO is not supported on this platform. Only Windows OSs are supported."
+      );
+      cy.ntlmSso(["nisse.com", "google.com"]);
     });
   });